增加二级评论功能；增加UITextView+Placeholder扩展；

+++ Weilefood/Helper/ServerHelper/WLServerHelper+Comment.h
@@ -22,7 +22,7 @@ typedef NS_ENUM(NSUInteger, WLCommentType);
  *  @param parentId 引用或回复评论ID
  *  @param callback
  */
-- (void)comment_addWithType:(WLCommentType)type refId:(NSUInteger)refId content:(NSString *)content parentId:(NSUInteger)parentId callback:(void (^)(WLApiInfoModel *apiInfo, NSError *error))callback;
+- (void)comment_addWithType:(WLCommentType)type refId:(long long)refId content:(NSString *)content parentId:(long long)parentId callback:(void (^)(WLApiInfoModel *apiInfo, NSError *error))callback;
 
 /**
  *  删除评论
@@ -30,7 +30,7 @@ typedef NS_ENUM(NSUInteger, WLCommentType);
  *  @param commentId 评论ID
  *  @param callback
  */
-- (void)comment_deleteWithCommentId:(NSUInteger)commentId callback:(void (^)(WLApiInfoModel *apiInfo, NSError *error))callback;
+- (void)comment_deleteWithCommentId:(long long)commentId callback:(void (^)(WLApiInfoModel *apiInfo, NSError *error))callback;
 
 /**
  *  获取某对象的评论列表。(NSArray<WLCommentModel>)apiResult
@@ -41,7 +41,7 @@ typedef NS_ENUM(NSUInteger, WLCommentType);
  *  @param pageSize 返回的最大记录数
  *  @param callback
  */
-- (void)comment_getListWithType:(WLCommentType)type refId:(NSUInteger)refId maxDate:(NSDate *)maxDate pageSize:(NSUInteger)pageSize callback:(void (^)(WLApiInfoModel *apiInfo, NSArray *apiResult, NSError *error))callback;
+- (void)comment_getListWithType:(WLCommentType)type refId:(long long)refId maxDate:(NSDate *)maxDate pageSize:(NSUInteger)pageSize callback:(void (^)(WLApiInfoModel *apiInfo, NSArray *apiResult, NSError *error))callback;
 
 /**
  *  获取我的评论列表。(NSArray<WLCommentModel>)apiResult
@@ -52,6 +52,6 @@ typedef NS_ENUM(NSUInteger, WLCommentType);
  *  @param pageSize 返回的最大记录数
  *  @param callback
  */
-- (void)comment_getMyListWithType:(WLCommentType)type refId:(NSUInteger)refId maxDate:(NSDate *)maxDate pageSize:(NSUInteger)pageSize callback:(void (^)(WLApiInfoModel *apiInfo, NSArray *apiResult, NSError *error))callback;
+- (void)comment_getMyListWithType:(WLCommentType)type refId:(long long)refId maxDate:(NSDate *)maxDate pageSize:(NSUInteger)pageSize callback:(void (^)(WLApiInfoModel *apiInfo, NSArray *apiResult, NSError *error))callback;
 
 @end
